pyproject.toml
[project]
name = "urlshortener"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.4",
    "django-debug-toolbar>=6.0.0",
    "django-extensions>=4.1",
    "django-phonenumber-field[phonenumbers]>=8.1.0",
    "django-query-analyzer>=0.0.17",
    "django-rest-framework>=0.1.0",
    "djangorestframework-simplejwt[crypto]>=5.5.1",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
    "sentry-sdk[django]>=2.33.2",
]
